refactor : 랭킹 시스템 코드리뷰 반영 (삭제상품 필터, 날짜 검증, payload 방어, lockValue 스코프)

+++ apps/commerce-streamer/src/main/java/com/loopers/application/ranking/RankingScoreService.java
@@ -71,7 +71,7 @@ public class RankingScoreService {
     public void addOrderScores(List<OrderItemPayload> items) {
         Map<Long, Double> aggregated = new HashMap<>();
         for (OrderItemPayload item : items) {
-            double rawValue = Math.max((long) item.price() * item.quantity(), 1);
+            double rawValue = Math.max((long) item.price() * (long) item.quantity(), 1);
             double score = orderWeight * Math.log10(rawValue);
             aggregated.merge(item.productId(), score, Double::sum);
         }
